requirements.txt
asgiref==3.8.1
astroid==3.3.5
dill==0.3.9
Django==5.1.2
django-stubs==5.1.1
django-stubs-ext==5.1.1
isort==5.13.2
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-decouple==3.8
sqlparse==0.5.1
tomli==2.0.2
tomlkit==0.13.2
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
